Add admin controllers and update views for cinema app
This commit introduces new controllers for managing branches, employees, rooms, showtimes, tickets, and users in the cinema application. Each controller is equipped with dependency injection for `ModelContext` and includes an `Index` action for data retrieval.

The corresponding `Index.cshtml` files have been updated with structured layouts and data bindings, featuring tables for displaying records with action buttons. The sidebar navigation has been enhanced to include links to the new controllers.

Additionally, the layout files have been restructured for consistency, and model updates enforce non-nullable relationships where necessary. CSS adjustments improve responsiveness and layout design.

These changes significantly enhance the functionality and user experience of the admin panel.

+++ DoAn/Models/Cinema/Room.cs
@@ -12,7 +12,7 @@ public partial class Room
 
     public string? Name { get; set; }
 
-    public short? Capacity { get; set; }
+    public int? Capacity { get; set; }
 
     public int? RoomTypeId { get; set; }
